@@ -8,123 +8,132 @@
     "url": "https://github.com/keystonejs/keystone.git"
   },
   "dependencies": {
-    "async": "2.5.0",
-    "asyncdi": "1.1.0",
-    "babel-core": "6.26.0",
-    "babel-plugin-transform-object-assign": "6.22.0",
-    "babel-polyfill": "6.26.0",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-react": "6.24.1",
-    "babel-preset-stage-2": "6.24.1",
-    "babelify": "7.3.0",
-    "babyparse": "0.4.6",
-    "bcrypt-nodejs": "~0.0.3",
-    "blacklist": "1.1.4",
-    "body-parser": "1.18.2",
-    "brfs": "1.4.3",
-    "browserify": "13.3.0",
-    "browserify-shim": "3.8.14",
-    "caller-id": "0.1.0",
-    "chalk": "1.1.3",
-    "classnames": "2.2.5",
-    "cloudinary": "1.9.0",
-    "cloudinary-microurl": "1.2.0",
-    "compression": "1.7.1",
-    "connect-flash": "0.1.1",
-    "cookie-parser": "1.4.3",
-    "debug": "2.6.8",
-    "display-name": "0.1.0",
+    "@babel/core": "^7.4.4",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/plugin-proposal-decorators": "^7.4.4",
+    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
+    "@babel/plugin-proposal-function-sent": "^7.2.0",
+    "@babel/plugin-proposal-json-strings": "^7.2.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.2.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.2.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/plugin-syntax-import-meta": "^7.2.0",
+    "@babel/plugin-transform-object-assign": "^7.2.0",
+    "@babel/polyfill": "^7.4.4",
+    "@babel/preset-env": "^7.4.4",
+    "@babel/preset-react": "^7.0.0",
+    "async": "^2.6.2",
+    "asyncdi": "^1.1.0",
+    "babelify": "^10.0.0",
+    "babyparse": "^0.4.6",
+    "bcrypt-nodejs": "^0.0.3",
+    "blacklist": "^1.1.4",
+    "body-parser": "^1.19.0",
+    "brfs": "^2.0.2",
+    "browserify": "^16.2.3",
+    "browserify-shim": "^3.8.14",
+    "caller-id": "^0.1.0",
+    "chalk": "^2.4.2",
+    "classnames": "^2.2.6",
+    "cloudinary": "^1.14.0",
+    "cloudinary-microurl": "^1.4.0",
+    "compression": "^1.7.4",
+    "connect-flash": "^0.1.1",
+    "cookie-parser": "^1.4.4",
+    "debug": "^4.1.1",
+    "display-name": "^0.1.0",
     "dumb-passwords": "^0.2.1",
-    "ejs": "2.5.6",
-    "elemental": "0.6.1",
-    "embedly": "2.1.0",
-    "errorhandler": "1.5.0",
-    "es6-promise": "4.1.1",
-    "express": "4.16.1",
-    "express-request-language": "1.1.12",
-    "express-session": "1.15.6",
-    "expression-match": "0.0.17",
-    "fs-extra": "1.0.0",
-    "glamor": "^2.20.25",
-    "grappling-hook": "3.0.0",
-    "i": "0.3.5",
-    "kerberos": "0.0.23",
-    "keystone-storage-namefunctions": "1.1.1",
-    "keystone-utils": "0.4.0",
-    "knox": "keystonejs/knox",
-    "less-middleware": "2.2.1",
-    "letsencrypt-express": "2.0.6",
-    "list-to-array": "1.1.0",
-    "lodash": "4.17.4",
-    "marked": "0.3.9",
-    "method-override": "2.3.10",
-    "mime-types": "2.1.15",
-    "moment": "2.18.1",
-    "mongoose": "4.9.2",
-    "morgan": "1.9.0",
-    "multer": "0.1.8",
-    "numeral": "2.0.6",
-    "object-assign": "4.1.1",
-    "qs": "4.0.0",
-    "queryfilter": "0.0.4",
-    "range_check": "1.4.0",
-    "react": "15.4.2",
-    "react-addons-css-transition-group": "15.6.2",
-    "react-alt-text": "2.0.0",
-    "react-color": "2.13.8",
-    "react-day-picker": "2.5.0",
-    "react-dnd": "2.5.3",
-    "react-dnd-html5-backend": "2.5.3",
-    "react-dom": "15.4.2",
-    "react-domify": "0.2.6",
-    "react-images": "0.5.6",
-    "react-markdown": "2.5.0",
-    "react-redux": "5.0.6",
-    "react-router": "3.0.2",
-    "react-router-redux": "4.0.8",
-    "react-select": "1.0.0-rc.10",
-    "redux": "3.7.2",
-    "redux-saga": "0.15.6",
-    "redux-thunk": "2.2.0",
-    "sanitize-filename": "1.6.1",
-    "sanitize-html": "^1.14.1",
-    "scmp": "1.0.2",
-    "semver": "5.4.1",
-    "serve-favicon": "2.4.5",
-    "string-to-stream": "1.1.0",
-    "vkey": "1.0.1",
-    "watchify": "3.9.0",
-    "xhr": "2.4.0"
+    "ejs": "^2.6.1",
+    "elemental": "^0.6.1",
+    "embedly": "^2.1.1",
+    "errorhandler": "^1.5.1",
+    "es6-promise": "^4.2.6",
+    "express": "^4.17.0",
+    "express-request-language": "^1.1.15",
+    "express-session": "^1.16.1",
+    "expression-match": "^0.0.17",
+    "fs-extra": "^8.0.1",
+    "glamor": "^2.20.40",
+    "grappling-hook": "^3.0.0",
+    "greenlock-express": "^2.7.8",
+    "i": "^0.3.6",
+    "keystone-storage-namefunctions": "^1.1.1",
+    "keystone-tinymce": "^0.1.0",
+    "keystone-utils": "^0.4.0",
+    "knox-s3": "^0.9.5",
+    "less-middleware": "^3.1.0",
+    "list-to-array": "^1.1.0",
+    "lodash": "^4.17.11",
+    "marked": "^0.6.2",
+    "method-override": "^3.0.0",
+    "mime-types": "^2.1.24",
+    "moment": "^2.24.0",
+    "mongoose": "^4.13.14",
+    "morgan": "^1.9.1",
+    "multer": "^1.4.1",
+    "numeral": "^2.0.6",
+    "object-assign": "^4.1.1",
+    "qs": "^6.7.0",
+    "queryfilter": "^0.0.4",
+    "range_check": "^1.4.0",
+    "react": "^15.4.2",
+    "react-addons-css-transition-group": "^15.6.2",
+    "react-alt-text": "^2.0.0",
+    "react-color": "^2.14.1",
+    "react-day-picker": "^2.5.0",
+    "react-dnd": "^2.5.3",
+    "react-dnd-html5-backend": "^2.5.3",
+    "react-dom": "^15.4.2",
+    "react-domify": "^0.2.6",
+    "react-images": "^0.5.6",
+    "react-markdown": "^2.5.0",
+    "react-redux": "^5.0.6",
+    "react-router": "^3.0.2",
+    "react-router-redux": "^4.0.8",
+    "redux": "^3.7.2",
+    "redux-saga": "^0.15.6",
+    "redux-thunk": "^2.2.0",
+    "react-select": "^1.2.4",
+    "sanitize-filename": "^1.6.1",
+    "tinymce": "^4.9.2",
+    "sanitize-html": "^1.20.1",
+    "scmp": "^1.0.2",
+    "semver": "^6.0.0",
+    "serve-favicon": "^2.5.0",
+    "string-to-stream": "^1.1.1",
+    "vkey": "^1.0.1",
+    "watchify": "^3.11.1",
+    "xhr": "^2.5.0"
   },
   "devDependencies": {
-    "@ljharb/eslint-config": "^11.0.0",
-    "browserify-middleware": "7.1.0",
-    "codeclimate-test-reporter": "0.5.0",
-    "connect-mongo": "1.3.2",
-    "core-assert": "0.2.1",
-    "cross-env": "3.1.4",
-    "disc": "1.3.3",
-    "enzyme": "2.8.2",
-    "eslint": "3.17.0",
-    "eslint-config-keystone": "3.0.0",
-    "eslint-config-keystone-react": "1.0.0",
-    "eslint-config-postcss": "^2.0.2",
-    "eslint-plugin-react": "6.9.0",
-    "istanbul": "0.4.5",
-    "keystone-email": "1.0.5",
-    "keystone-nightwatch-e2e": "0.2.14",
-    "mocha": "3.5.3",
-    "must": "0.13.4",
-    "proxyquire": "1.8.0",
-    "react-addons-test-utils": "15.6.2",
-    "react-engine": "4.3.0",
-    "rimraf": "2.6.2",
-    "sinon": "2.3.2",
-    "superagent": "3.6.0",
-    "supertest": "3.0.0",
-    "uglify-js": "2.8.18",
-    "updtr": "1.0.0",
-    "watch": "1.0.2"
+    "@babel/register": "^7.4.4",
+    "browserify-middleware": "^8.1.1",
+    "codeclimate-test-reporter": "^0.5.1",
+    "connect-mongo": "^2.0.3",
+    "core-assert": "^1.0.0",
+    "cross-env": "^5.2.0",
+    "disc": "^1.3.3",
+    "enzyme": "^2.9.1",
+    "eslint": "^5.16.0",
+    "eslint-config-keystone": "^3.0.0",
+    "eslint-config-keystone-react": "^1.0.0",
+    "eslint-config-postcss": "^3.0.7",
+    "eslint-plugin-react": "^7.13.0",
+    "istanbul": "^0.4.5",
+    "keystone-email": "^1.1.0",
+    "keystone-nightwatch-e2e": "^0.2.14",
+    "mocha": "^3.5.3",
+    "must": "^0.13.4",
+    "proxyquire": "^2.1.0",
+    "react-addons-test-utils": "^15.6.2",
+    "react-engine": "^4.5.1",
+    "rimraf": "^2.6.3",
+    "sinon": "^2.4.1",
+    "superagent": "^5.0.5",
+    "supertest": "^3.4.2",
+    "uglify-js": "^3.5.13",
+    "updtr": "^3.1.0",
+    "watch": "^1.0.2"
   },
   "browserify": {
     "transform": [
