@@ -91,7 +91,8 @@ var PrimaryNavigation = React.createClass({
 
 		return this.props.sections.map((section) => {
 			// Get the link and the class name
-			const href = section.lists[0].external ? section.lists[0].path : `${Keystone.adminPath}/${section.lists[0].path}`;
+			const to = !section.lists[0].external && `${Keystone.adminPath}/${section.lists[0].path}`;
+			const href = section.lists[0].external && section.lists[0].path;
 			const isActive = this.props.currentSectionKey && this.props.currentSectionKey === section.key;
 			const className = isActive ? 'primary-navbar__item--active' : null;
 
@@ -101,7 +102,8 @@ var PrimaryNavigation = React.createClass({
 					key={section.key}
 					label={section.label}
 					className={className}
-					to={href}
+					to={to}
+					href={href}
 				>
 					{section.label}
 				</PrimaryNavItem>
