@@ -68,7 +68,7 @@ export function * evalQueryParams () {
 	const { cachedQuery } = yield select(state => state.active);
 	const { currentList } = yield select(state => state.lists);
 
-	if (pathname !== `/keystone/${currentList.id}`) return;
+	if (pathname !== `${Keystone.adminPath}/${currentList.id}`) return;
 
 	if (isEqual(query, cachedQuery)) {
 		yield put({ type: actions.QUERY_HAS_NOT_CHANGED });
