@@ -59,7 +59,9 @@ module.exports = {
 		var updatedValues = this.state.values;
 		var updateIndex = updatedValues.indexOf(i);
 		var newValue = event.value || event.target.value;
-		updatedValues[updateIndex].value = this.cleanInput ? this.cleanInput(newValue) : newValue;
+		if (this.isValid === undefined || this.isValid(newValue)) {
+			updatedValues[updateIndex].value = this.cleanInput ? this.cleanInput(newValue) : newValue;
+		}
 		this.setState({
 			values: updatedValues,
 		});
