@@ -36,7 +36,9 @@ boolean.prototype.validateInput = function (data, callback) {
 
 boolean.prototype.validateRequiredInput = function (item, data, callback) {
 	var value = this.getValueFromData(data);
-	var result = (value && value !== 'false') || typeof item.get(this.path) === 'boolean' ? true : false;
+	var result = value && value !== 'false'
+		? true
+		: false;
 	utils.defer(callback, result);
 };
 
