@@ -287,13 +287,18 @@ module.exports = Field.create({
 		);
 	},
 
+	// This renders a hidden input that holds the payload data for how the field
+	// should be updated. It should be upload:{filename}, undefined, or 'remove'
 	renderActionInput () {
 		if (!this.shouldRenderField()) return null;
 
 		if (this.state.userSelectedFile || this.state.removeExisting) {
-			const value = this.state.userSelectedFile
-				? `upload:${this.state.uploadFieldPath}`
-				: '';
+			let value = '';
+			if (this.state.userSelectedFile) {
+				value = `upload:${this.state.uploadFieldPath}`;
+			} else if (this.state.removeExisting && this.props.autoCleanup) {
+				value = 'delete';
+			}
 			return (
 				<input
 					name={this.getInputName(this.props.path)}
