@@ -9,6 +9,7 @@ import {
 	InlineGroupSection as Section,
 } from '../../../admin/client/App/elemental';
 import transparentSwatch from './transparent-swatch';
+import coloredSwatch from './colored-swatch';
 import theme from '../../../admin/client/theme';
 
 const ColorField = Field.create({
@@ -61,7 +62,8 @@ const ColorField = Field.create({
 		return (this.props.value) ? (
 			<span
 				className={className}
-				style={{ backgroundColor: this.props.value }}
+				style={{ color: this.props.value }}
+				dangerouslySetInnerHTML={{ __html: coloredSwatch }}
 			/>
 		) : (
 			<span
@@ -71,6 +73,7 @@ const ColorField = Field.create({
 		);
 	},
 	renderField () {
+
 		const { displayColorPicker } = this.state;
 
 		return (
@@ -86,7 +89,7 @@ const ColorField = Field.create({
 						/>
 					</Section>
 					<Section>
-						<Button onClick={this.handleClick} cssStyles={classes.button} data-e2e-type-color__button>
+						<Button onClick={this.handleClick} style={classes.button} data-e2e-type-color__button>
 							{this.renderSwatch()}
 						</Button>
 					</Section>
@@ -115,13 +118,13 @@ const ColorField = Field.create({
 /* eslint quote-props: ["error", "as-needed"] */
 const classes = {
 	button: {
-		background: 'white',
+		background: 'white !important',
 		padding: 4,
 		width: theme.component.height,
 
-		':hover': {
-			background: 'white',
-		},
+		// ':hover': {
+		// 	background: 'white',
+		// },
 	},
 	blockout: {
 		bottom: 0,
@@ -135,14 +138,15 @@ const classes = {
 		marginTop: 10,
 		position: 'absolute',
 		left: 0,
-		zIndex: 2,
+		zIndex: 500,
 	},
 	swatch: {
 		borderRadius: 1,
-		boxShadow: 'inset 0 0 0 1px rgba(0,0,0,0.1)',
+		boxShadow: '0 0 0 1px rgba(0,0,0,0.1)',
 		display: 'block',
-		height: '100%',
-		width: '100%',
+		' svg': {
+			display: 'block',
+		},
 	},
 };
 
