@@ -28,6 +28,7 @@ module.exports = Field.create({
 		note: React.PropTypes.string,
 		onChange: React.PropTypes.func,
 		path: React.PropTypes.string,
+		todayButton: React.PropTypes.bool,
 		value: React.PropTypes.string,
 	},
 
@@ -85,9 +86,12 @@ module.exports = Field.create({
 						value={value}
 					/>
 				</Section>
-				<Section>
-					<Button onClick={this.setToday}>Today</Button>
-				</Section>
+				{
+					this.props.todayButton
+					&& <Section>
+						<Button onClick={this.setToday}>Today</Button>
+					</Section>
+				}
 			</Group>
 		);
 	},
