@@ -11,9 +11,9 @@ var NumberColumn = React.createClass({
 	},
 	renderValue () {
 		const value = this.props.data.fields[this.props.col.path];
-		if (!value || isNaN(value)) return null;
+		if (value === undefined || isNaN(value)) return null;
 
-		const formattedValue = (this.props.col.path === 'money') ? numeral(value).format('$0,0.00') : value;
+		const formattedValue = (this.props.col.type === 'money') ? numeral(value).format('$0,0.00') : value;
 
 		return formattedValue;
 	},
