@@ -22,7 +22,7 @@ var moduleRoot = (function (_rootPath) {
  * Keystone Class
  */
 var Keystone = function () {
-	grappling.mixin(this).allowHooks('pre:static', 'pre:bodyparser', 'pre:session', 'pre:logger', 'pre:admin', 'pre:routes', 'pre:render', 'updates', 'signin', 'signout');
+	grappling.mixin(this).allowHooks('pre:static', 'pre:bodyparser', 'pre:session', 'pre:logger', 'pre:admin', 'pre:adminroutes', 'pre:routes', 'pre:render', 'updates', 'signin', 'signout');
 	this.lists = {};
 	this.fieldTypes = {};
 	this.paths = {};
@@ -38,6 +38,7 @@ var Keystone = function () {
 		'module root': moduleRoot,
 		'frame guard': 'sameorigin',
 		'cache admin bundles': true,
+		'handle uploads': true,
 	};
 	this._redirects = {};
 
