@@ -32,6 +32,7 @@ module.exports = function openDatabaseConnection (callback) {
 				rs_name: replicaData.db.replicaSetOptions.rs_name,
 				readPreference: replicaData.db.replicaSetOptions.readPreference,
 			},
+			useMongoClient: true,
 		};
 
 		debug('connecting to replica set');
@@ -40,7 +41,9 @@ module.exports = function openDatabaseConnection (callback) {
 	} else {
 		debug('connecting to mongo');
 		keystone.initDatabaseConfig();
-		keystone.mongoose.connect(keystone.get('mongo'), keystone.get('mongo options'));
+		var mongo_options = keystone.get('mongo options') || {};
+		mongo_options.useMongoClient = true;
+		keystone.mongoose.connect(keystone.get('mongo'), mongo_options);
 	}
 
 	keystone.mongoose.connection.on('error', function (err) {
