@@ -55,7 +55,7 @@ function start (events) {
 
 		var ssl = keystone.get('ssl');
 		var unixSocket = keystone.get('unix socket');
-		var startupMessages = ['KeystoneJS Started:'];
+		var startupMessages = [`KeystoneJS v${keystone.version} started:`];
 
 		async.parallel([
 			// HTTP Server
