@@ -1,6 +1,6 @@
-var _ = require('lodash');
-var keystone = require('../../');
-var Types = require('../fieldTypes');
+var _ = require("lodash");
+var keystone = require("../../");
+var Types = require("../fieldTypes");
 
 /**
  * List track option
@@ -9,11 +9,10 @@ var Types = require('../fieldTypes');
  * as well as the user who created/updated it.
  */
 
-module.exports = function track () {
-
+module.exports = function track() {
 	var list = this;
-	var options = list.get('track');
-	var userModel = keystone.get('user model');
+	var options = list.get("track");
+	var userModel = keystone.get("user model");
 
 	if (!options) {
 		// if the track setting is falsy, bail
@@ -24,15 +23,19 @@ module.exports = function track () {
 		createdAt: false,
 		createdBy: false,
 		updatedAt: false,
-		updatedBy: false,
+		updatedBy: false
 	};
 	var fields = {};
 
 	// ensure track is a boolean or an object
 	if (!_.isBoolean(options) && !_.isObject(options)) {
-		throw new Error('Invalid List "track" option for ' + list.key + '\n'
-			+ '"track" must be a boolean or an object.\n\n'
-			+ 'See http://keystonejs.com/docs/database/#lists-options for more information.');
+		throw new Error(
+			'Invalid List "track" option for ' +
+				list.key +
+				"\n" +
+				'"track" must be a boolean or an object.\n\n' +
+				"See http://v4.keystonejs.com/docs/database/#lists-options for more information."
+		);
 	}
 
 	if (_.isBoolean(options)) {
@@ -41,12 +44,17 @@ module.exports = function track () {
 			createdAt: true,
 			createdBy: true,
 			updatedAt: true,
-			updatedBy: true,
+			updatedBy: true
 		};
 	}
 
 	// if all track fields are set to false, then there's nothing to track
-	if (!options.createdAt && !options.createdBy && !options.updatedAt && !options.updatedBy) {
+	if (
+		!options.createdAt &&
+		!options.createdBy &&
+		!options.updatedAt &&
+		!options.updatedBy
+	) {
 		return;
 	}
 
@@ -54,24 +62,34 @@ module.exports = function track () {
 	options = _.extend({}, defaultOptions, options);
 
 	// validate option fields
-	_.forEach(options, function (value, key) {
-
+	_.forEach(options, function(value, key) {
 		var fieldName;
 
 		// make sure the key isn't already defined as a field
 		if (_.has(list.fields, key)) {
-			throw new Error('Invalid List "track" option for ' + list.key + '\n'
-				+ '"' + key + '" is already defined in the Schema.');
+			throw new Error(
+				'Invalid List "track" option for ' +
+					list.key +
+					"\n" +
+					'"' +
+					key +
+					'" is already defined in the Schema.'
+			);
 		}
 
 		// make sure it's a valid track option field
 		if (_.has(defaultOptions, key)) {
-
 			// make sure the option field value is either a boolean or a string
-			if (!_.isBoolean(value) && typeof value !== 'string') {
-				throw new Error('Invalid List "track" option for ' + list.key + '\n'
-					+ '"' + key + '" must be a boolean or a string.\n\n'
-					+ 'See http://keystonejs.com/docs/database/#lists-options for more information.');
+			if (!_.isBoolean(value) && typeof value !== "string") {
+				throw new Error(
+					'Invalid List "track" option for ' +
+						list.key +
+						"\n" +
+						'"' +
+						key +
+						'" must be a boolean or a string.\n\n' +
+						"See http://v4.keystonejs.com/docs/database/#lists-options for more information."
+				);
 			}
 
 			if (value) {
@@ -81,33 +99,46 @@ module.exports = function track () {
 				list.map(key, fieldName);
 
 				switch (key) {
-					case 'createdAt':
-					case 'updatedAt':
-						fields[fieldName] = { type: Date, noedit: true, collapse: true, index: true };
+					case "createdAt":
+					case "updatedAt":
+						fields[fieldName] = {
+							type: Date,
+							noedit: true,
+							collapse: true,
+							index: true
+						};
 						break;
 
-					case 'createdBy':
-					case 'updatedBy':
-						fields[fieldName] = { type: Types.Relationship, ref: userModel, noedit: true, collapse: true, index: true };
+					case "createdBy":
+					case "updatedBy":
+						fields[fieldName] = {
+							type: Types.Relationship,
+							ref: userModel,
+							noedit: true,
+							collapse: true,
+							index: true
+						};
 						break;
 				}
 			}
 		} else {
-			throw new Error('Invalid List "track" option for ' + list.key + '\n'
-				+ 'valid field options are "createdAt", "createdBy", "updatedAt", an "updatedBy".\n\n'
-				+ 'See http://keystonejs.com/docs/database/#lists-options for more information.');
+			throw new Error(
+				'Invalid List "track" option for ' +
+					list.key +
+					"\n" +
+					'valid field options are "createdAt", "createdBy", "updatedAt", an "updatedBy".\n\n' +
+					"See http://v4.keystonejs.com/docs/database/#lists-options for more information."
+			);
 		}
-
 	});
 
 	// add track fields to schema
-	list.add('Meta', fields);
+	list.add("Meta", fields);
 
 	list.tracking = options;
 
 	// add the pre-save schema plugin
-	list.schema.pre('save', function (next) {
-
+	list.schema.pre("save", function(next) {
 		var now = new Date();
 
 		// set createdAt/createdBy on new docs
@@ -131,7 +162,5 @@ module.exports = function track () {
 		}
 
 		next();
-
 	});
-
 };
