@@ -49,7 +49,7 @@ function signinWithUser (user, req, res, onSuccess) {
 	req.session.regenerate(function () {
 		req.user = user;
 		req.session.userId = user.id;
-		// if the user has a password set, store a persistence cookie to resume sessions
+		// if the user has a password set, store a persistent cookie to resume sessions
 		if (keystone.get('cookie signin') && user.password) {
 			var userToken = user.id + ':' + hash(user.password);
 			var cookieOpts = _.defaults({}, keystone.get('cookie signin options'), {
@@ -57,6 +57,9 @@ function signinWithUser (user, req, res, onSuccess) {
 				httpOnly: true,
 				maxAge: 10 * 24 * 60 * 60 * 1000,
 			});
+			if (req.session.cookie) {
+				req.session.cookie.maxAge = cookieOpts.maxAge;
+			}
 			res.cookie('keystone.uid', userToken, cookieOpts);
 		}
 		onSuccess(user);
