@@ -94,7 +94,7 @@ FSAdapter.prototype.uploadFile = function (file, callback) {
 	var options = this.options;
 	this.getFilename(file, function (err, filename) {
 		if (err) return callback(err);
-		filename = sanitize(filename);
+		filename = sanitize(filename) + path.parse(file.originalname).ext;
 		debug('Uploading file with filename: %s', filename);
 		var uploadPath = path.resolve(options.path, filename);
 		var fsOptions = {};
