@@ -2,7 +2,7 @@ var mongoose = require('mongoose');
 var mongoUri = 'mongodb://localhost/test';
 
 function dropTestDatabase(done) {
-	mongoose.connect(mongoUri,function (err) {
+	mongoose.connect(mongoUri, { useMongoClient: true }, function (err) {
 		if (!err) {
 			mongoose.connection.db.dropDatabase(function (err) {
 				mongoose.connection.close(function (err) {
