@@ -35,15 +35,22 @@ describe('Test dependsOn and required', function () {
 				done(error);
 			}
 
+			// suppressing console log output
+			const backupLog = console.error;
+			console.error = () => null;
+
 			var newPost = new DependsOn.model({
 				title: 'new post',
-				state: 'published'
-			});
-			newPost.save(function (err) {
-				demand(err).be.a.object();
-				done();
+				state: 'published',
+				publishedDate: undefined,
 			});
 
+			newPost.save(function (err) {
+				demand(err).be.a.object();
+
+				console.error = backupLog;
+				done();
+			});
 		});
 
 	});
